Implement events in runner to track the execution process
